Reveal rect annotations from anno.file.
